Feat: `AllListingsReport` deserializing
A toolbox full of report sources prevent entries requiring user-input.
The "All Listings Report" holds the most complete item listing
information. Importantly, items that have not yet been sent to FBA are
likely to be described within the listings report.

This commit adds a custom deserialzer for translating AMZ's condition
encoding as well. Currently, functionality for the translation is small,
but it will likely become prevalent in the future. See the `Condition` enum for
it's variants.