Feat: retrieve trunk names from directory path
Implementation for this feature is a simple `read_dir` call, converting
the `OsStr` into an owned `String`, and finally filtering out anything
that is not a Json file.

Down the line, there will likely be reasoning to include other
serialized Json files alongside entries. An extra call to filter can be
used with a fallible conversion to yank anything out when needed.