Feat: display trunks component. Config declaration
The potential of Plaine comes from allowing the user to see as much
information as possible, while only showing what they need to see.
Displaying trunks, and being able to swap between them is somehow
a luxury we have never had.

Not a single tool that we have used for "planning shipments" has ever
displayed this information while also showing the contents of whichever
shipment you are on.

Someone should have the ability to manage multiple shipments at a single
time, and to manage something, it helps to know that it exists.