Feat: branch status uses most recent entries
Previously, sorting statuses was implemented through checking for
partial equality on the `Status` enum. This only works if you never have
the need to return back to an old status.

When an item needs to be shifted off of the branch after a confirmation,
you need to reopen it back to the a different status.